Move various structs into separate file
